style: renamed TokenFactory for semantic
